Separate out the caching mechanism from the original image source
Initially the resized images were written back to the same source as
they were fetched from. However this limits us to sources that are both
readable and writable. The store is now separated in a ImageSource and a
cache. The resizer will also read the cache now whereas previously we
relied on an external caching mechanism such as varnish or nginx.