fix issues found by test.check.
One issue was found that too small sizes resulted in an exception. This
exception is now converted to a 400 error. The Resampling needs a
minimal image of 3x3.

Another issue was that a cmyk image could not be saved to jpeg. So now
we convert it to a rgb first.